requirements.txt
Werkzeug==0.16.0
Flask==1.0.2
Flask-Login==0.4.1
Flask-WTF==0.14.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2